+++ src/pages/Cart/index.tsx
@@ -8,7 +8,7 @@ import PageTitle from "../../Components/PageTitle";
 import SubmitBox from "../../Components/SubmitBox";
 import { COLOR } from "../../constants/theme";
 
-import CartItem from "../../Containers/CartItem";
+import CartItem from "./CartItem";
 import { FlexCenter } from "../../SharedStyled/Flex";
 import { RootState } from "../../store";
 import { Container, Main, AllDealControlBox, Section, AllDealSelect, AllDealDelete, CartListTitle } from "./styles";
